data
Previously, styler parameters were "resolved" before being injected into the
template. This feature was designed to accomodate the use of reactive data
values, but has turned out to be unnecessary and difficult to understand.

Now, stylers are passed their parameters as plain old data. This has resulted
in an interface that is smaller, simpler, and less weird.